Revealing modifier item overlay helmets with Revealing have an overlay on the item sprite now. it is a modified version of the conarm Traveller's Goggles item overlay sprite.
TraitDieInstantly now uses int instead of double for the random value.

moved dependency defintions from build.gradle to dependencies.gradle. less bothersome to find there.

+++ src/main/java/com/emperdog/tinkertantrum/trait/TraitDieInstantly.java
@@ -21,7 +21,7 @@ public class TraitDieInstantly extends AbstractTrait {
     @Override
     public void afterBlockBreak(ItemStack tool, World world, IBlockState state, BlockPos pos, EntityLivingBase player, boolean wasEffective) {
         if(!world.isRemote){
-            double random = world.rand.nextInt(Integer.MAX_VALUE);
+            int random = world.rand.nextInt(Integer.MAX_VALUE);
             //TinkerTantrumMod.LOGGER.info(random);
             if(random == Integer.MAX_VALUE - 1)
                 dieInstantly(player);
